+++ lib/dropdown_textfield.dart
@@ -5,10 +5,53 @@ import 'package:dropdown_textfield/tooltip_widget.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
+class IconProperty {
+  final IconData? icon;
+  final Color? color;
+  final double? size;
+  IconProperty({this.icon, this.color, this.size});
+}
+
+class CheckBoxProperty {
+  final MouseCursor? mouseCursor;
+  final Color? activeColor;
+  final MaterialStateProperty<Color?>? fillColor;
+  final Color? checkColor;
+  final bool tristate;
+  final MaterialTapTargetSize? materialTapTargetSize;
+  final VisualDensity? visualDensity;
+  final Color? focusColor;
+  final Color? hoverColor;
+  final MaterialStateProperty<Color?>? overlayColor;
+  final double? splashRadius;
+  final FocusNode? focusNode;
+  final bool autofocus;
+  final OutlinedBorder? shape;
+  final BorderSide? side;
+  static const double width = 18.0;
+  CheckBoxProperty({
+    this.tristate = false,
+    this.mouseCursor,
+    this.activeColor,
+    this.fillColor,
+    this.checkColor,
+    this.focusColor,
+    this.hoverColor,
+    this.overlayColor,
+    this.splashRadius,
+    this.materialTapTargetSize,
+    this.visualDensity,
+    this.focusNode,
+    this.autofocus = false,
+    this.shape,
+    this.side,
+  });
+}
+
 class DropDownTextField extends StatefulWidget {
   const DropDownTextField(
       {Key? key,
-      this.singleController,
+      this.controller,
       this.initialValue,
       required this.dropDownList,
       this.padding,
@@ -20,54 +63,72 @@ class DropDownTextField extends StatefulWidget {
       this.readOnly = true,
       this.dropdownRadius = 12,
       this.textFieldDecoration,
+      this.dropDownIconProperty,
       this.dropDownItemCount = 6,
       this.searchFocusNode,
       this.textFieldFocusNode,
       this.searchAutofocus = false,
+      this.searchDecoration,
       this.searchShowCursor,
       this.searchKeyboardType,
       this.listSpace = 0,
       this.clearOption = true,
+      this.clearIconProperty,
       // this.keyboardHeight = 450,
       this.listPadding,
       this.listTextStyle})
-      : assert(!(initialValue != null && singleController != null),
-            "you cannot add both initialValue and singleController,\nset initial value using controller \n\tEg: SingleValueDropDownController(data:initial value) "),
+      : assert(
+          !(initialValue != null && controller != null),
+          "you cannot add both initialValue and singleController,\nset initial value using controller \n\tEg: SingleValueDropDownController(data:initial value) ",
+        ),
+        assert(
+          !(controller != null &&
+              !(controller is SingleValueDropDownController)),
+          "controller must be type of SingleValueDropDownController",
+        ),
+        checkBoxProperty = null,
         isMultiSelection = false,
-        isForceMultiSelectionClear = false,
-        displayCompleteItem = false,
+        singleController = controller,
         multiController = null,
+        displayCompleteItem = false,
         submitButtonColor = null,
         submitButtonText = null,
         submitButtonTextStyle = null,
         super(key: key);
-  const DropDownTextField.multiSelection(
-      {Key? key,
-      this.multiController,
-      this.displayCompleteItem = false,
-      this.initialValue,
-      required this.dropDownList,
-      this.padding,
-      this.textStyle,
-      @Deprecated('Use multiController instead. Will be removed in next version')
-          this.isForceMultiSelectionClear = false,
-      this.onChanged,
-      this.validator,
-      this.isEnabled = true,
-      this.dropdownRadius = 12,
-      this.textFieldDecoration,
-      this.dropDownItemCount = 6,
-      this.searchFocusNode,
-      this.textFieldFocusNode,
-      this.listSpace = 0,
-      this.clearOption = true,
-      this.submitButtonColor,
-      this.submitButtonText,
-      this.submitButtonTextStyle,
-      this.listPadding,
-      this.listTextStyle})
-      : assert(initialValue == null || multiController == null,
+  const DropDownTextField.multiSelection({
+    Key? key,
+    this.controller,
+    this.displayCompleteItem = false,
+    this.initialValue,
+    required this.dropDownList,
+    this.padding,
+    this.textStyle,
+    this.onChanged,
+    this.validator,
+    this.isEnabled = true,
+    this.dropdownRadius = 12,
+    this.dropDownIconProperty,
+    this.textFieldDecoration,
+    this.dropDownItemCount = 6,
+    this.searchFocusNode,
+    this.textFieldFocusNode,
+    this.listSpace = 0,
+    this.clearOption = true,
+    this.clearIconProperty,
+    this.submitButtonColor,
+    this.submitButtonText,
+    this.submitButtonTextStyle,
+    this.listPadding,
+    this.listTextStyle,
+    this.checkBoxProperty,
+  })  : assert(initialValue == null || controller == null,
             "you cannot add both initialValue and multiController\nset initial value using controller\n\tMultiValueDropDownController(data:initial value)"),
+        assert(
+          !(controller != null &&
+              !(controller is MultiValueDropDownController)),
+          "controller must be type of MultiValueDropDownController",
+        ),
+        multiController = controller,
         isMultiSelection = true,
         enableSearch = false,
         readOnly = true,
@@ -75,8 +136,13 @@ class DropDownTextField extends StatefulWidget {
         searchKeyboardType = null,
         searchShowCursor = null,
         singleController = null,
+        searchDecoration = null,
         // keyboardHeight = 0,
         super(key: key);
+
+  ///single and multiple dropdown controller.
+  ///It must be type of SingleValueDropDownController or MultiValueDropDownController.
+  final dynamic controller;
 
   ///single dropdown controller,
   final SingleValueDropDownController? singleController;
@@ -105,11 +171,11 @@ class DropDownTextField extends StatefulWidget {
 
   final EdgeInsets? padding;
 
-  ///by setting isForceMultiSelectionClear=true to deselect selected item,only applicable for multi selection dropdown
-  final bool isForceMultiSelectionClear;
-
   ///override default textfield decoration
   final InputDecoration? textFieldDecoration;
+
+  ///customize dropdown icon size and color
+  final IconProperty? dropDownIconProperty;
 
   ///by setting isEnabled=false to disable textfield,default value true
   final bool isEnabled;
@@ -130,6 +196,9 @@ class DropDownTextField extends StatefulWidget {
   final FocusNode? searchFocusNode;
   final FocusNode? textFieldFocusNode;
 
+  ///override default search decoration
+  final InputDecoration? searchDecoration;
+
   ///override default search keyboard type,only applicable if enableSearch=true,
   final TextInputType? searchKeyboardType;
 
@@ -142,6 +211,9 @@ class DropDownTextField extends StatefulWidget {
 
   ///by set clearOption=false to hide clear suffix icon button from textfield.
   final bool clearOption;
+
+  ///customize Clear icon size and color
+  final IconProperty? clearIconProperty;
 
   ///space between textfield and list ,default value is 0
   final double listSpace;
@@ -160,6 +232,9 @@ class DropDownTextField extends StatefulWidget {
 
   ///dropdown list item text style
   final TextStyle? listTextStyle;
+
+  ///customize checkbox property
+  final CheckBoxProperty? checkBoxProperty;
 
   @override
   _DropDownTextFieldState createState() => _DropDownTextFieldState();
@@ -245,7 +320,7 @@ class _DropDownTextFieldState extends State<DropDownTextField>
           var index = _dropDownList.indexWhere((element) =>
               element.name.trim() == widget.initialValue[i].trim());
           if (index != -1) {
-            _multiSelectionValue[i] = true;
+            _multiSelectionValue[index] = true;
           }
         }
         int count =
@@ -291,14 +366,14 @@ class _DropDownTextFieldState extends State<DropDownTextField>
             _multiSelectionValue.add(false);
           }
         }
-        if (widget.isForceMultiSelectionClear &&
-            _multiSelectionValue.isNotEmpty) {
-          _multiSelectionValue = [];
-          _cnt.text = "";
-          for (int i = 0; i < _dropDownList.length; i++) {
-            _multiSelectionValue.add(false);
-          }
-        }
+        // if (widget.isForceMultiSelectionClear &&
+        //     _multiSelectionValue.isNotEmpty) {
+        //   _multiSelectionValue = [];
+        //   _cnt.text = "";
+        //   for (int i = 0; i < _dropDownList.length; i++) {
+        //     _multiSelectionValue.add(false);
+        //   }
+        // }
 
         // if (widget.multiController != null) {
         //   List<DropDownValueModel> multiCnt = [];
@@ -450,14 +525,19 @@ class _DropDownTextFieldState extends State<DropDownTextField>
             decoration: widget.textFieldDecoration != null
                 ? widget.textFieldDecoration!.copyWith(
                     suffixIcon: (_cnt.text.isEmpty || !widget.clearOption)
-                        ? const Icon(
-                            Icons.arrow_drop_down_outlined,
+                        ? Icon(
+                            widget.dropDownIconProperty?.icon ??
+                                Icons.arrow_drop_down_outlined,
+                            size: widget.dropDownIconProperty?.size,
+                            color: widget.dropDownIconProperty?.color,
                           )
                         : widget.clearOption
                             ? InkWell(
                                 onTap: clearFun,
-                                child: const Icon(
-                                  Icons.clear,
+                                child: Icon(
+                                  widget.clearIconProperty?.icon ?? Icons.clear,
+                                  size: widget.clearIconProperty?.size,
+                                  color: widget.clearIconProperty?.color,
                                 ),
                               )
                             : null,
@@ -467,14 +547,19 @@ class _DropDownTextFieldState extends State<DropDownTextField>
                     hintText: _hintText,
                     hintStyle: const TextStyle(fontWeight: FontWeight.normal),
                     suffixIcon: (_cnt.text.isEmpty || !widget.clearOption)
-                        ? const Icon(
-                            Icons.arrow_drop_down_outlined,
+                        ? Icon(
+                            widget.dropDownIconProperty?.icon ??
+                                Icons.arrow_drop_down_outlined,
+                            size: widget.dropDownIconProperty?.size,
+                            color: widget.dropDownIconProperty?.color,
                           )
                         : widget.clearOption
                             ? InkWell(
                                 onTap: clearFun,
-                                child: const Icon(
-                                  Icons.clear,
+                                child: Icon(
+                                  widget.clearIconProperty?.icon ?? Icons.clear,
+                                  size: widget.clearIconProperty?.size,
+                                  color: widget.clearIconProperty?.color,
                                 ),
                               )
                             : null,
@@ -688,6 +773,7 @@ class _DropDownTextFieldState extends State<DropDownTextField>
                       searchHeight: _searchWidgetHeight,
                       searchKeyboardType: widget.searchKeyboardType,
                       searchAutofocus: _searchAutofocus,
+                      searchDecoration: widget.searchDecoration,
                       searchShowCursor: widget.searchShowCursor,
                       listPadding: _listPadding,
                       onSearchTap: () {
@@ -703,7 +789,9 @@ class _DropDownTextFieldState extends State<DropDownTextField>
                         if (_isScrollPadding) {
                           shiftOverlayEntry2to1();
                         }
-                      })
+                      },
+                      clearIconProperty: widget.clearIconProperty,
+                    )
                   : MultiSelection(
                       buttonTextStyle: widget.submitButtonTextStyle,
                       buttonText: widget.submitButtonText,
@@ -747,6 +835,7 @@ class _DropDownTextFieldState extends State<DropDownTextField>
 
                         setState(() {});
                       },
+                      checkBoxProperty: widget.checkBoxProperty,
                     ),
             ),
           ),
@@ -775,7 +864,9 @@ class SingleSelection extends StatefulWidget {
       this.onSearchTap,
       this.onSearchSubmit,
       this.listTextStyle,
-      required this.listPadding})
+      this.searchDecoration,
+      required this.listPadding,
+      this.clearIconProperty})
       : super(key: key);
   final List<DropDownValueModel> dropDownList;
   final ValueSetter onChanged;
@@ -794,6 +885,8 @@ class SingleSelection extends StatefulWidget {
   final Function? onSearchSubmit;
   final TextStyle? listTextStyle;
   final ListPadding listPadding;
+  final InputDecoration? searchDecoration;
+  final IconProperty? clearIconProperty;
 
   @override
   State<SingleSelection> createState() => _SingleSelectionState();
@@ -803,6 +896,7 @@ class _SingleSelectionState extends State<SingleSelection> {
   late List<DropDownValueModel> newDropDownList;
   late TextEditingController _searchCnt;
   late FocusScopeNode _focusScopeNode;
+  late InputDecoration _inpDec;
   onItemChanged(String value) {
     setState(() {
       if (value.isEmpty) {
@@ -819,6 +913,7 @@ class _SingleSelectionState extends State<SingleSelection> {
   @override
   void initState() {
     _focusScopeNode = FocusScopeNode();
+    _inpDec = widget.searchDecoration ?? InputDecoration();
     if (widget.searchAutofocus) {
       widget.searchFocusNode.requestFocus();
     }
@@ -862,8 +957,8 @@ class _SingleSelectionState extends State<SingleSelection> {
                     widget.onSearchTap!();
                   }
                 },
-                decoration: InputDecoration(
-                  hintText: 'Search Here...',
+                decoration: _inpDec.copyWith(
+                  hintText: _inpDec.hintText ?? 'Search Here...',
                   suffixIcon: GestureDetector(
                     onTap: () {
                       widget.mainFocusNode.requestFocus();
@@ -871,8 +966,12 @@ class _SingleSelectionState extends State<SingleSelection> {
                       onItemChanged("");
                     },
                     child: widget.searchFocusNode.hasFocus
-                        ? const InkWell(
-                            child: Icon(Icons.close),
+                        ? InkWell(
+                            child: Icon(
+                              widget.clearIconProperty?.icon ?? Icons.close,
+                              size: widget.clearIconProperty?.size,
+                              color: widget.clearIconProperty?.color,
+                            ),
                           )
                         : const SizedBox.shrink(),
                   ),
@@ -936,7 +1035,8 @@ class MultiSelection extends StatefulWidget {
       this.buttonTextStyle,
       required this.listTileHeight,
       required this.listPadding,
-      this.listTextStyle})
+      this.listTextStyle,
+      this.checkBoxProperty})
       : super(key: key);
   final List<DropDownValueModel> dropDownList;
   final ValueSetter onChanged;
@@ -948,6 +1048,7 @@ class MultiSelection extends StatefulWidget {
   final double listTileHeight;
   final TextStyle? listTextStyle;
   final ListPadding listPadding;
+  final CheckBoxProperty? checkBoxProperty;
 
   @override
   _MultiSelectionState createState() => _MultiSelectionState();
@@ -1013,6 +1114,25 @@ class _MultiSelectionState extends State<MultiSelection> {
                                 });
                               }
                             },
+                            tristate:
+                                widget.checkBoxProperty?.tristate ?? false,
+                            mouseCursor: widget.checkBoxProperty?.mouseCursor,
+                            activeColor: widget.checkBoxProperty?.activeColor,
+                            fillColor: widget.checkBoxProperty?.fillColor,
+                            checkColor: widget.checkBoxProperty?.checkColor,
+                            focusColor: widget.checkBoxProperty?.focusColor,
+                            hoverColor: widget.checkBoxProperty?.hoverColor,
+                            overlayColor: widget.checkBoxProperty?.overlayColor,
+                            splashRadius: widget.checkBoxProperty?.splashRadius,
+                            materialTapTargetSize:
+                                widget.checkBoxProperty?.materialTapTargetSize,
+                            visualDensity:
+                                widget.checkBoxProperty?.visualDensity,
+                            focusNode: widget.checkBoxProperty?.focusNode,
+                            autofocus:
+                                widget.checkBoxProperty?.autofocus ?? false,
+                            shape: widget.checkBoxProperty?.shape,
+                            side: widget.checkBoxProperty?.side,
                           ),
                         ],
                       ),